refactor(capabilities): make audio own its kinds and split audio.rs
ADR-0121 audio slice: pure code movement, no behavior change. The
~5520-line audio.rs becomes an audio/ submodule and the 11
aether.audio.* mail kinds move out of aether-kinds into audio/kinds.rs,
riding the always-on audio marker so wasm guests addressing the cap
still see the types. Kind ids and wire shapes are unchanged (a kind id
is fnv1a_64 over name+schema, stable across a crate move).

- audio.rs renamed to audio/mod.rs; mod native carved along its
  cohesion seams into native-only siblings config/event/schedule/
  instrument/voice/sample/track/synth, cfg-gated like decode/sfz. The
  bridge mod native keeps AudioCapability + its actor block and the
  inline test module.
- audio/kinds.rs holds NoteOn/NoteOff/SetMasterGain plus the control
  plane (SetMasterGainResult, the schedule batch, the track lane, the
  sampled-bank loader) with re-homed NAME, round-trip, and
  descriptor-inventory tests.
- aether-capabilities gains inventory (non-wasm target dep, for the Kind
  derive's native submit) and bytemuck with derive via the audio marker
  feature so the Pod cast-kinds compile on the marker-only path.
- The moved declarations and their tests are deleted from aether-kinds;
  headless chassis re-points its SetMasterGain import at the cap.

Closes #2155
