+++ crates/aether-capabilities/src/audio/mod.rs
@@ -53,12 +53,20 @@
 //! `NoteOn` / `NoteOff` are dropped silently and `SetMasterGain`
 //! replies `Err` so agents fail fast instead of hanging.
 
+// ADR-0121: the `aether.audio.*` mail vocabulary the cap owns, riding
+// the always-on `audio` marker (not native-gated) so a wasm guest
+// addressing the cap through the marker feature sees the types. The
+// glob re-export surfaces every audio kind at `aether_capabilities::audio::*`
+// for external callers.
+pub mod kinds;
+pub use kinds::*;
+
 // Handler-signature kinds must be importable at file root because
 // `#[bridge]` emits `impl HandlesKind<K> for X {}` markers as siblings
-// of the mod (always-on, outside the cfg gate).
-use aether_kinds::{
-    LoadInstrument, NoteOff, NoteOn, PlayTrack, ReadResult, Schedule, SetMasterGain, StopTrack,
-};
+// of the mod (always-on, outside the cfg gate). The audio kinds resolve
+// through the `kinds` glob above; `ReadResult` is an `aether.fs` kind (a
+// different cap) the audio cap receives as the track-load reply.
+use aether_kinds::ReadResult;
 
 // `AudioConfig` rides through file root for chassis-bin consumers
 // that build it from env (`from_env`) and pass it to
@@ -67,7 +75,7 @@ use aether_kinds::{
 // the config struct (sends are typed; config is the chassis's
 // concern).
 #[cfg(all(not(target_arch = "wasm32"), feature = "audio-native"))]
-pub use native::{AudioConfig, AudioConfigLayer, AudioOverlay};
+pub use config::{AudioConfig, AudioConfigLayer, AudioOverlay};
 
 // ADR-0103 §1 decode/resample core (`crates/aether-capabilities/src/audio/decode.rs`).
 // Native-only — it pulls `hound` and `std`; the marker-only `audio`
@@ -83,1959 +91,95 @@ mod decode;
 #[cfg(all(not(target_arch = "wasm32"), feature = "audio-native"))]
 mod sfz;
 
+// ADR-0121 cohesion submodules carved out of the bridge `mod native`. Each
+// is native-only (gated like `decode` / `sfz`); `mod native` reaches them
+// via `super::`. The seams: config (the derive-Config layer), event (the
+// cpal event queue), schedule (the ADR-0104 heap entry), instrument (the
+// built-in registry), voice (the synthesis kernels), sample (the ADR-0103
+// sampled banks), track (the ADR-0103 mixer lane), and synth (the mixer
+// aggregate + cpal pipeline build).
+#[cfg(all(not(target_arch = "wasm32"), feature = "audio-native"))]
+mod config;
+#[cfg(all(not(target_arch = "wasm32"), feature = "audio-native"))]
+mod event;
+#[cfg(all(not(target_arch = "wasm32"), feature = "audio-native"))]
+mod instrument;
+#[cfg(all(not(target_arch = "wasm32"), feature = "audio-native"))]
+mod sample;
+#[cfg(all(not(target_arch = "wasm32"), feature = "audio-native"))]
+mod schedule;
+#[cfg(all(not(target_arch = "wasm32"), feature = "audio-native"))]
+mod synth;
+#[cfg(all(not(target_arch = "wasm32"), feature = "audio-native"))]
+mod track;
+#[cfg(all(not(target_arch = "wasm32"), feature = "audio-native"))]
+mod voice;
+
 #[aether_actor::bridge(singleton, feature = "audio-native")]
 mod native {
-    use std::cmp::{Ordering, Reverse};
-    use std::collections::BinaryHeap;
     use std::collections::HashMap;
     use std::collections::VecDeque;
-    use std::f32::consts::TAU;
-    use std::io::Cursor;
     use std::str::from_utf8;
     use std::sync::Arc;
     use std::sync::mpsc;
     use std::thread::{self, JoinHandle};
 
-    use cpal::traits::{DeviceTrait, HostTrait, StreamTrait};
-    use crossbeam_queue::ArrayQueue;
-
+    use super::kinds::{
+        LoadInstrument, LoadInstrumentResult, PlayTrack, PlayTrackResult, Schedule, ScheduleResult,
+        SetMasterGainResult, StopTrack,
+    };
     use aether_actor::{Manual, OutboundReply, actor};
     use aether_data::{MailboxId, Source, SourceAddr};
-    use aether_kinds::{
-        LoadInstrument, LoadInstrumentResult, PlayTrack, PlayTrackResult, Read, ReadResult,
-        Schedule, ScheduleResult, ScheduledEvent, ScheduledNote, SetMasterGainResult, StopTrack,
-    };
+    use aether_kinds::{Read, ReadResult};
 
     use aether_substrate::actor::native::{NativeActor, NativeCtx, NativeInitCtx, TaskDone};
     use aether_substrate::chassis::error::BootError;
 
     use crate::fs::FsCapability;
 
-    use super::decode::{DecodeError, decode_wav_to_mono};
-    use super::sfz::{SfzLoop, SfzRegion, parse_sfz};
+    use super::decode::decode_wav_to_mono;
+    use super::sfz::parse_sfz;
     use super::{NoteOff, NoteOn, SetMasterGain};
 
-    /// Linear fade-out duration (seconds) applied when a track is stopped,
-    /// so `stop_track` releases through a short ramp instead of truncating
-    /// to a click (ADR-0103 §3).
-    const TRACK_FADE_SECS: f32 = 0.005;
-    // confique consumes `parse_flag` through `#[config(parse_env = …)]`;
-    // IntelliJ-Rust doesn't trace macro-attr path args (Qodana FP), but
-    // rustc + clippy do.
-    #[allow(unused_imports)]
-    use crate::config_env::parse_flag;
-    use core::fmt;
-
-    /// Capacity of the event queue between the cap's handlers and the
-    /// audio-callback consumer. 1024 slots hold ~10 seconds of a dense
-    /// 100-note-per-second stream; overflow is warn-dropped, which the
-    /// ADR-0039 timing-quantization section already documents as a v1
-    /// limitation (tight-burst percussion may drop notes under load).
-    const EVENT_QUEUE_CAPACITY: usize = 1024;
-
-    /// Maximum concurrent voices before voice-stealing kicks in. Chosen
-    /// as "more than a string section fits in one component" — on
-    /// saturation, voice-steal always evicts the oldest sounding note,
-    /// never causing audio glitches.
-    const MAX_VOICES: usize = 64;
-
-    /// Maximum note events one `aether.audio.schedule` batch may carry
-    /// (ADR-0104). A batch crosses the event queue as a single slot, so
-    /// this bounds the synth's pending heap rather than the queue; 8192
-    /// events is several minutes of a dense melody (note-on + note-off per
-    /// note). An over-cap batch rejects atomically in the handler reply.
-    const SCHEDULE_MAX_EVENTS: usize = 8192;
-
-    /// Furthest future a scheduled event may be parked, in milliseconds
-    /// (ADR-0104). The horizon bounds how much future a sender can hold in
-    /// the pending heap; ten minutes is generous for a tune dispatched in
-    /// one call. An over-horizon `at_millis` rejects the whole batch.
-    const SCHEDULE_MAX_MILLIS: u32 = 600_000;
-
-    /// Resolved configuration for the audio synth. Chassis mains read
-    /// env vars (`AETHER_AUDIO_DISABLE`, `AETHER_AUDIO_SAMPLE_RATE`)
-    /// into an `AudioConfig` and pass it to `with_actor::<AudioCapability>(cfg)`
-    /// (issue 464). Tests build an `AudioConfig` directly.
-    ///
-    /// ADR-0090 unit g (iamacoffeepot/aether#1264): the
-    /// `#[derive(aether_substrate::Config)]` emits the env-shaped
-    /// `AudioConfigLayer`, the clap-shaped `AudioOverlay`, the
-    /// `FromArgvThenEnv` impl, and the inherent `from_env` /
-    /// `from_argv_then_env` shims. `requested_sample_rate`'s type
-    /// `Option<u32>` triggers the macro's type-driven
-    /// `Option<numeric>` shape: the Layer holds `Option<String>` and
-    /// `from_layer` does the soft `.parse().ok()` so an unparseable
-    /// value lands as `None` (indistinguishable from unset, matching
-    /// the prior reader).
-    #[derive(Clone, Debug, Default, aether_substrate::Config)]
-    #[config(env_prefix = "AETHER_AUDIO", cli_prefix = "audio")]
-    pub struct AudioConfig {
-        /// `AETHER_AUDIO_DISABLE=1` skips cpal init entirely. The cap
-        /// still claims its mailbox and replies `Err` to `SetMasterGain`
-        /// so agents fail fast instead of hanging. `env` + `cli_long`
-        /// overrides pin the historical wire shape (no `D` suffix on
-        /// `DISABLE`; `--audio-disable` not `--audio-disabled`).
-        #[config(
-            env = "AETHER_AUDIO_DISABLE",
-            cli_long = "audio-disable",
-            default = false,
-            parse = parse_flag
-        )]
-        pub disabled: bool,
-        /// `AETHER_AUDIO_SAMPLE_RATE=<hz>` requests a specific rate. If
-        /// the device doesn't support it, boot falls back to nop
-        /// (ADR-0039 — non-fatal). `layer_field = "sample_rate"` drops
-        /// the `requested_` prefix on the Layer / env / CLI side so the
-        /// historical names are unchanged.
-        #[config(layer_field = "sample_rate", env = "AETHER_AUDIO_SAMPLE_RATE")]
-        pub requested_sample_rate: Option<u32>,
-    }
-
-    /// Event a handler pushes into the audio callback's queue. The
-    /// `sender_mailbox` is baked in here (not re-derived on the callback
-    /// side) so the callback stays branch-minimal.
-    ///
-    /// Not `Copy`: the track-start variant carries an `Arc`'d PCM buffer
-    /// (the decoded asset) and owned namespace / path strings (ADR-0103
-    /// §3). The queue never required `Copy`.
-    #[derive(Clone, Debug)]
-    enum AudioEvent {
-        NoteOn {
-            sender_mailbox: MailboxId,
-            pitch: u8,
-            velocity: u8,
-            instrument_id: u8,
-        },
-        NoteOff {
-            sender_mailbox: MailboxId,
-            pitch: u8,
-            instrument_id: u8,
-        },
-        SetMasterGain {
-            gain: f32,
-        },
-        /// Start (or restart) a track in the dedicated mixer lane. `pcm`
-        /// is already mono and resampled to the device rate, so the
-        /// callback walks it by index. Keyed by `(sender_mailbox, lane,
-        /// namespace, path)` — re-sending the same key restarts the track.
-        TrackStart {
-            sender_mailbox: MailboxId,
-            lane: Option<String>,
-            namespace: String,
-            path: String,
-            pcm: Arc<[f32]>,
-            gain: f32,
-            looping: bool,
-        },
-        /// Fade out and retire the track at this key. A no-op if no track
-        /// matches (matching `note_off`).
-        TrackStop {
-            sender_mailbox: MailboxId,
-            lane: Option<String>,
-            namespace: String,
-            path: String,
-        },
-        /// Append a loaded sampled-instrument bank to the synth's registry
-        /// (ADR-0103 §4). The cap assigns `id` from `BUILTINS.len()` upward
-        /// in load order and the synth pushes the bank in receipt order, so
-        /// the two stay in lockstep — a `note_on` whose `instrument_id`
-        /// walks past the built-ins indexes this table. `bank` is the
-        /// assembled, device-rate PCM bank behind an `Arc` (shared with the
-        /// voices it spawns).
-        RegisterInstrument {
-            id: u8,
-            bank: Arc<SampleBank>,
-        },
-        /// A validated batch of timed note events (ADR-0104). `sender_mailbox`
-        /// is the scheduling sender, baked in so every scheduled note keys
-        /// its voice (and note-off matching) by the original caller. The
-        /// synth converts each event's `at_millis` to an absolute due frame
-        /// against its frame clock at the instant it drains this event, so
-        /// the whole batch shares one receipt timebase and chords stay
-        /// aligned. One queue slot carries the entire tune.
-        Schedule {
-            sender_mailbox: MailboxId,
-            events: Vec<ScheduledEvent>,
-        },
-    }
-
-    /// One pending scheduled note in the synth's min-heap (ADR-0104).
-    /// Ordered by `(due_frame, seq)` only — `seq` is a monotonic stamp in
-    /// batch-arrival order, so events that fall on the same frame fire in
-    /// the order they were sent. The note payload takes no part in
-    /// ordering, which is why the ordering impls are hand-written rather
-    /// than derived (`ScheduledNote` is not `Ord`).
-    struct ScheduledEntry {
-        due_frame: u64,
-        seq: u64,
-        sender_mailbox: MailboxId,
-        note: ScheduledNote,
-    }
-
-    impl PartialEq for ScheduledEntry {
-        fn eq(&self, other: &Self) -> bool {
-            (self.due_frame, self.seq) == (other.due_frame, other.seq)
-        }
-    }
-
-    impl Eq for ScheduledEntry {}
-
-    impl PartialOrd for ScheduledEntry {
-        fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-            Some(self.cmp(other))
-        }
-    }
-
-    impl Ord for ScheduledEntry {
-        fn cmp(&self, other: &Self) -> Ordering {
-            (self.due_frame, self.seq).cmp(&(other.due_frame, other.seq))
-        }
-    }
-
-    /// Convert a play-at offset in milliseconds to a frame count at the
-    /// device rate (ADR-0104). Added to the frame clock at receipt to land
-    /// the absolute due frame.
-    fn millis_to_frames(at_millis: u32, sample_rate: f32) -> u64 {
-        // `at_millis` is bounded by `SCHEDULE_MAX_MILLIS` and the device
-        // rate is a small positive integer, so the product is well within
-        // u64 and non-negative.
-        #[allow(
-            clippy::cast_possible_truncation,
-            clippy::cast_sign_loss,
-            clippy::cast_precision_loss
-        )]
-        let frames = (f64::from(at_millis) / 1000.0 * f64::from(sample_rate)) as u64;
-        frames
-    }
-
-    /// Producer side of the audio event queue. The cap holds one (after
-    /// building the pipeline) and pushes events on every inbound `NoteOn`
-    /// / `NoteOff` / `SetMasterGain`.
-    #[derive(Clone)]
-    struct AudioEventSender {
-        queue: Arc<ArrayQueue<AudioEvent>>,
-    }
-
-    impl AudioEventSender {
-        fn push(&self, event: AudioEvent) -> Result<(), AudioEvent> {
-            self.queue.push(event)
-        }
-    }
-
-    fn new_event_channel() -> (AudioEventSender, Arc<ArrayQueue<AudioEvent>>) {
-        let queue = Arc::new(ArrayQueue::new(EVENT_QUEUE_CAPACITY));
-        (
-            AudioEventSender {
-                queue: Arc::clone(&queue),
-            },
-            queue,
-        )
-    }
-
-    /// Primitive waveform the oscillator shapes. `Saw` is a downward ramp
-    /// scaled to ±1; `Pluck` reuses `Saw` geometry but pairs with a
-    /// fast-decay envelope — kept implicit by the patch table.
-    ///
-    /// `Noise` is the percussion source: white noise from a per-voice
-    /// xorshift32 PRNG (seeded from the voice key, so a fixed key renders
-    /// the same sequence every run), shaped by a one-pole lowpass whose
-    /// `lowpass` coefficient is a patch constant — `1.0` passes the raw
-    /// white noise (bright, a hat), a smaller value smooths it (darker, a
-    /// snare body). `tone_mix` blends in a fixed-level sine at the voice's
-    /// base frequency under the noise (`0.0` is pure noise); it is the one
-    /// patch field that turns a hat patch into a snare.
-    #[derive(Copy, Clone, Debug)]
-    enum Wave {
-        Sine,
-        Square,
-        Triangle,
-        Saw,
-        Noise { lowpass: f32, tone_mix: f32 },
-    }
-
-    /// Envelope shape — linear segments at sample-rate resolution. Values
-    /// are held in seconds; the voice converts to per-sample step on
-    /// instantiation so the hot loop is add-only.
-    #[derive(Copy, Clone, Debug)]
-    struct Adsr {
-        attack_s: f32,
-        decay_s: f32,
-        sustain: f32,
-        release_s: f32,
-    }
-
-    /// Optional per-patch pitch envelope on the oscillator kernel — the
-    /// whole identity of a kick. The voice's phase step is multiplied by
-    /// a ratio that starts at `start_ratio` and decays exponentially
-    /// toward `1.0` (the note's base frequency) with the given time
-    /// constant. The decay is precomputed as a per-sample multiplier at
-    /// `note_on`, so the hot loop pays one extra multiply.
-    #[derive(Copy, Clone, Debug)]
-    struct PitchSweep {
-        /// Phase-step multiplier at the note's onset. `4.0` starts two
-        /// octaves above the base frequency; `1.0` is no sweep.
-        start_ratio: f32,
-        /// Exponential time constant (seconds) of the fall back to the
-        /// base frequency. Short (tens of millis) for a punchy kick.
-        time_constant_secs: f32,
-    }
-
-    /// Number of sine partials in a partial-bank voice. Fixed so the
-    /// voice stays `Copy` and stack-friendly in the pool; the hot loop is
-    /// one `sin`, one multiply-accumulate, and one decay multiply per
-    /// partial.
-    const PARTIAL_COUNT: usize = 8;
-
-    /// Reference pitch (MIDI C4) for partial-bank decay scaling. A note's
-    /// per-partial decay rates scale by `f0 / REFERENCE_FREQ`, so higher
-    /// notes ring shorter and lower notes longer.
-    const REFERENCE_FREQ: f32 = 261.625_57;
-
-    /// Relative amplitude below which a sustaining (un-released)
-    /// partial-bank voice frees itself. Piano partials decay
-    /// exponentially and never reach exactly zero, so the voice retires
-    /// once its summed partial energy crosses this floor.
-    const PARTIAL_SILENCE_FLOOR: f32 = 1.0e-4;
-
-    /// A struck/sustained partial-bank voice patch. Partial `n` is tuned
-    /// to `n * f0 * sqrt(1 + inharmonicity * n^2)` plus a small
-    /// per-partial detune; `partial_amps` is the spectral shape (tilted
-    /// toward upper partials by velocity via `brightness_tilt`); each
-    /// partial decays at `decay_base * (1 + i * decay_spread)` scaled by
-    /// pitch. A global attack/release ramp wraps the bank.
-    #[derive(Copy, Clone, Debug)]
-    struct PartialBankDef {
-        /// Stiffness coefficient `B`: stretches overtone `n` to
-        /// `n * f0 * sqrt(1 + B * n^2)`. `0.0` is perfectly harmonic.
-        inharmonicity: f32,
-        /// Per-partial base amplitude (the spectral shape). Normalised at
-        /// `note_on` so overall level comes from velocity, not this sum.
-        partial_amps: [f32; PARTIAL_COUNT],
-        /// Fundamental decay rate (per second) at the reference pitch.
-        /// `0.0` sustains indefinitely (the pad).
-        decay_base: f32,
-        /// Per-partial-index decay multiplier: partial `i` decays at
-        /// `decay_base * (1 + i * decay_spread)`, so upper partials fade
-        /// first (a string's brightness dropping as it rings).
-        decay_spread: f32,
-        /// Per-partial detune fraction. Alternating ± across partials
-        /// gives the slow beating of a multi-string course.
-        detune: f32,
-        /// Velocity-to-brightness tilt. Higher velocity multiplies the
-        /// upper partials' share so a harder strike reads brighter.
-        brightness_tilt: f32,
-        /// Global attack ramp (seconds). Near-zero for a struck string,
-        /// long for the pad's slow swell.
-        attack_s: f32,
-        /// Global release ramp (seconds) on `note_off` — the damper.
-        release_s: f32,
-    }
-
-    /// Voice kernel a patch selects. The five original patches stay
-    /// `Oscillator`; the partial-bank patches add struck-string and
-    /// sustained timbres without a wire change.
-    #[derive(Copy, Clone, Debug)]
-    enum VoiceDef {
-        Oscillator { wave: Wave, adsr: Adsr },
-        PartialBank(PartialBankDef),
-    }
-
-    /// Full instrument patch. Agents address instruments by numeric id
-    /// into the built-in registry; the registry hands the voice a copy of
-    /// this struct at `note_on` time so each voice is self-contained.
-    #[derive(Copy, Clone, Debug)]
-    struct InstrumentDef {
-        name: &'static str,
-        voice: VoiceDef,
-        base_amp: f32,
-        /// Optional pitch envelope. Applies only to the `Oscillator`
-        /// kernel (the partial bank ignores it); `None` is the common
-        /// case. `Some` is the falling-frequency thump of a kick.
-        pitch_sweep: Option<PitchSweep>,
-    }
-
-    /// The v1 instrument registry. Index matches `NoteOn.instrument_id`.
-    /// Reordering these is a breaking change on the wire — adds go at
-    /// the end. Future follow-up: mailed patch definitions fill in past
-    /// the built-ins (ADR-0039 "runtime-defined patches" parked item).
-    const BUILTINS: &[InstrumentDef] = &[
-        InstrumentDef {
-            name: "sine_lead",
-            voice: VoiceDef::Oscillator {
-                wave: Wave::Sine,
-                adsr: Adsr {
-                    attack_s: 0.01,
-                    decay_s: 0.08,
-                    sustain: 0.7,
-                    release_s: 0.18,
-                },
-            },
-            base_amp: 0.35,
-            pitch_sweep: None,
-        },
-        InstrumentDef {
-            name: "square_bass",
-            voice: VoiceDef::Oscillator {
-                wave: Wave::Square,
-                adsr: Adsr {
-                    attack_s: 0.005,
-                    decay_s: 0.12,
-                    sustain: 0.6,
-                    release_s: 0.12,
-                },
-            },
-            base_amp: 0.22,
-            pitch_sweep: None,
-        },
-        InstrumentDef {
-            name: "triangle",
-            voice: VoiceDef::Oscillator {
-                wave: Wave::Triangle,
-                adsr: Adsr {
-                    attack_s: 0.02,
-                    decay_s: 0.1,
-                    sustain: 0.7,
-                    release_s: 0.2,
-                },
-            },
-            base_amp: 0.32,
-            pitch_sweep: None,
-        },
-        InstrumentDef {
-            name: "saw_lead",
-            voice: VoiceDef::Oscillator {
-                wave: Wave::Saw,
-                adsr: Adsr {
-                    attack_s: 0.01,
-                    decay_s: 0.15,
-                    sustain: 0.55,
-                    release_s: 0.15,
-                },
-            },
-            base_amp: 0.2,
-            pitch_sweep: None,
-        },
-        InstrumentDef {
-            name: "pluck",
-            voice: VoiceDef::Oscillator {
-                wave: Wave::Saw,
-                adsr: Adsr {
-                    attack_s: 0.002,
-                    decay_s: 0.35,
-                    sustain: 0.0,
-                    release_s: 0.05,
-                },
-            },
-            base_amp: 0.3,
-            pitch_sweep: None,
-        },
-        // id 5: struck-string piano. Slightly stretched partials, a
-        // bright-to-mellow decay (upper partials fade first), and a fast
-        // damper release on note_off.
-        InstrumentDef {
-            name: "piano",
-            voice: VoiceDef::PartialBank(PartialBankDef {
-                inharmonicity: 0.000_4,
-                partial_amps: [1.0, 0.6, 0.4, 0.25, 0.18, 0.12, 0.08, 0.05],
-                decay_base: 3.0,
-                decay_spread: 0.6,
-                detune: 0.000_8,
-                brightness_tilt: 0.5,
-                attack_s: 0.002,
-                release_s: 0.15,
-            }),
-            base_amp: 0.3,
-            pitch_sweep: None,
-        },
-        // id 6: electric piano. Same partial-bank shape, more inharmonic
-        // (bell-like), faster decay, and a brighter velocity response —
-        // a pure patch-table entry, no new machinery.
-        InstrumentDef {
-            name: "electric_piano",
-            voice: VoiceDef::PartialBank(PartialBankDef {
-                inharmonicity: 0.001,
-                partial_amps: [1.0, 0.3, 0.5, 0.2, 0.3, 0.15, 0.1, 0.06],
-                decay_base: 4.0,
-                decay_spread: 0.4,
-                detune: 0.001_2,
-                brightness_tilt: 0.7,
-                attack_s: 0.003,
-                release_s: 0.1,
-            }),
-            base_amp: 0.28,
-            pitch_sweep: None,
-        },
-        // id 7: slow-swell pad. Harmonic partials, a long attack, near-
-        // zero partial decay so it sustains while held, and a long
-        // release — the warm sustained bed no oscillator patch can do.
-        InstrumentDef {
-            name: "pad",
-            voice: VoiceDef::PartialBank(PartialBankDef {
-                inharmonicity: 0.0,
-                partial_amps: [1.0, 0.7, 0.5, 0.4, 0.3, 0.25, 0.2, 0.15],
-                decay_base: 0.0,
-                decay_spread: 0.0,
-                detune: 0.000_6,
-                brightness_tilt: 0.25,
-                attack_s: 0.8,
-                release_s: 0.6,
-            }),
-            base_amp: 0.18,
-            pitch_sweep: None,
-        },
-        // id 8: kick. A sine swept down from two octaves above the base
-        // frequency with a fast (30 ms) time constant, through a punchy
-        // no-sustain ADSR — the falling thump that defines a kick. `pitch`
-        // scales the base, so one patch covers kick through toms.
-        InstrumentDef {
-            name: "kick",
-            voice: VoiceDef::Oscillator {
-                wave: Wave::Sine,
-                adsr: Adsr {
-                    attack_s: 0.001,
-                    decay_s: 0.18,
-                    sustain: 0.0,
-                    release_s: 0.02,
-                },
-            },
-            base_amp: 0.9,
-            pitch_sweep: Some(PitchSweep {
-                start_ratio: 4.0,
-                time_constant_secs: 0.03,
-            }),
-        },
-        // id 9: hat. A short burst of bright (near-unfiltered) noise
-        // through a fast no-sustain ADSR. `pitch` shifts the register so
-        // one patch covers closed-versus-open flavours.
-        InstrumentDef {
-            name: "hat",
-            voice: VoiceDef::Oscillator {
-                wave: Wave::Noise {
-                    lowpass: 0.9,
-                    tone_mix: 0.0,
-                },
-                adsr: Adsr {
-                    attack_s: 0.001,
-                    decay_s: 0.04,
-                    sustain: 0.0,
-                    release_s: 0.02,
-                },
-            },
-            base_amp: 0.4,
-            pitch_sweep: None,
-        },
-        // id 10: snare. Darker (lowpassed) noise with a fixed-level sine
-        // body mixed under it (`tone_mix`) — the one patch field that
-        // separates a snare from a hat — through a short no-sustain ADSR.
-        InstrumentDef {
-            name: "snare",
-            voice: VoiceDef::Oscillator {
-                wave: Wave::Noise {
-                    lowpass: 0.5,
-                    tone_mix: 0.25,
-                },
-                adsr: Adsr {
-                    attack_s: 0.001,
-                    decay_s: 0.12,
-                    sustain: 0.0,
-                    release_s: 0.03,
-                },
-            },
-            base_amp: 0.5,
-            pitch_sweep: None,
-        },
-    ];
-
-    fn instrument_by_id(id: u8) -> Option<&'static InstrumentDef> {
-        BUILTINS.get(id as usize)
-    }
-
-    /// Number of built-in instruments. Used by the boot log so MCP
-    /// agents can cross-reference.
-    pub fn builtin_count() -> usize {
-        BUILTINS.len()
-    }
-
-    /// Names of the built-in instruments, in id order.
-    pub fn builtin_names() -> Vec<&'static str> {
-        BUILTINS.iter().map(|d| d.name).collect()
-    }
-
-    /// Envelope state machine. `Release` captures the level it was at
-    /// when the note was released, since a note can be released mid-attack
-    /// or mid-decay — the release ramp starts from that value, not from
-    /// the sustain level.
-    #[derive(Copy, Clone, Debug)]
-    enum EnvelopeStage {
-        Attack { t: f32 },
-        Decay { t: f32 },
-        Sustain,
-        Release { t: f32, from_level: f32 },
-        Done,
-    }
-
-    /// The oscillator voice kernel — a periodic waveform through a linear
-    /// ADSR. Every field is touched per sample, so the struct stays
-    /// compact for cache friendliness in the voice pool.
-    #[derive(Copy, Clone, Debug)]
-    struct OscVoice {
-        /// Oscillator phase in turns (`[0.0, 1.0)`), incremented by
-        /// `freq / sample_rate` per sample.
-        phase: f32,
-        /// Turns-per-sample step — precomputed so the per-sample path is
-        /// add-only.
-        phase_step: f32,
-        /// Base amplitude after velocity scaling; envelope multiplies this.
-        amplitude: f32,
-        wave: Wave,
-        adsr: Adsr,
-        envelope: EnvelopeStage,
-        /// xorshift32 PRNG state for the `Noise` wave, seeded from the
-        /// voice key. Unused (but harmless) for the periodic waves.
-        rng: u32,
-        /// One-pole lowpass memory for the `Noise` wave (the previous
-        /// filtered output).
-        lp_prev: f32,
-        /// Current pitch-sweep offset added to `1.0` to scale `phase_step`
-        /// this sample. `0.0` when the patch has no sweep.
-        sweep_offset: f32,
-        /// Per-sample multiplier the sweep offset decays by. `1.0` (no
-        /// decay) when the patch has no sweep — the offset is then `0.0`,
-        /// so the ratio stays `1.0`.
-        sweep_decay: f32,
-    }
-
-    /// One step of an xorshift32 PRNG, mapped to white noise in `[-1.0,
-    /// 1.0)`. The state is per-voice so percussion voices are independent
-    /// and a fixed seed is reproducible.
-    fn next_noise(state: &mut u32) -> f32 {
-        let mut x = *state;
-        x ^= x << 13;
-        x ^= x >> 17;
-        x ^= x << 5;
-        *state = x;
-        // Map the full u32 range to [-1.0, 1.0). The mantissa rounding is
-        // inaudible and irrelevant to a noise source.
-        #[allow(clippy::cast_precision_loss)]
-        let frac = (x as f32) / (u32::MAX as f32);
-        frac.mul_add(2.0, -1.0)
-    }
-
-    /// Seed the per-voice noise PRNG from the voice key
-    /// (`sender_mailbox`, `instrument_id`, `pitch`) so a fixed key renders
-    /// the same noise sequence every run. Forced non-zero — xorshift32 is
-    /// stuck at zero.
-    fn voice_seed(sender_mailbox: MailboxId, instrument_id: u8, pitch: u8) -> u32 {
-        // Truncating the 64-bit mailbox id into the hash is intended; the
-        // seed only needs to vary per key, not round-trip.
-        #[allow(clippy::cast_possible_truncation)]
-        let lo = sender_mailbox.0 as u32;
-        #[allow(clippy::cast_possible_truncation)]
-        let hi = (sender_mailbox.0 >> 32) as u32;
-        let mixed = lo.wrapping_mul(2_654_435_761)
-            ^ hi.wrapping_mul(40_503)
-            ^ u32::from(instrument_id).wrapping_mul(2_246_822_519)
-            ^ u32::from(pitch).wrapping_mul(3_266_489_917);
-        mixed | 1
-    }
-
-    impl OscVoice {
-        fn new(
-            pitch: u8,
-            velocity: u8,
-            wave: Wave,
-            adsr: Adsr,
-            base_amp: f32,
-            sample_rate: f32,
-            seed: u32,
-        ) -> Self {
-            let freq = 440.0 * ((f32::from(pitch) - 69.0) / 12.0).exp2();
-            let phase_step = freq / sample_rate;
-            let v = f32::from(velocity) / 127.0;
-            let amplitude = base_amp * v * v;
-            Self {
-                phase: 0.0,
-                phase_step,
-                amplitude,
-                wave,
-                adsr,
-                envelope: EnvelopeStage::Attack { t: 0.0 },
-                rng: seed,
-                lp_prev: 0.0,
-                sweep_offset: 0.0,
-                sweep_decay: 1.0,
-            }
-        }
-
-        /// Arm a pitch sweep on a freshly built voice. The offset starts
-        /// at `start_ratio - 1.0` and decays by `sweep_decay` per sample;
-        /// `next_sample` reads `1.0 + sweep_offset` as the phase-step
-        /// multiplier. A non-positive time constant is treated as no
-        /// sweep (the voice keeps its base frequency).
-        fn with_pitch_sweep(mut self, sweep: PitchSweep, sample_rate: f32) -> Self {
-            if sweep.time_constant_secs > 0.0 {
-                let dt = 1.0 / sample_rate;
-                self.sweep_offset = sweep.start_ratio - 1.0;
-                self.sweep_decay = (-dt / sweep.time_constant_secs).exp();
-            }
-            self
-        }
-
-        fn note_off(&mut self) {
-            let from_level = match self.envelope {
-                EnvelopeStage::Attack { t } => {
-                    if self.adsr.attack_s > 0.0 {
-                        (t / self.adsr.attack_s).clamp(0.0, 1.0)
-                    } else {
-                        1.0
-                    }
-                }
-                EnvelopeStage::Decay { t } => {
-                    if self.adsr.decay_s > 0.0 {
-                        let fall = (1.0 - self.adsr.sustain).mul_add(-(t / self.adsr.decay_s), 1.0);
-                        fall.clamp(self.adsr.sustain.min(1.0), 1.0)
-                    } else {
-                        self.adsr.sustain
-                    }
-                }
-                EnvelopeStage::Sustain => self.adsr.sustain,
-                EnvelopeStage::Release { .. } | EnvelopeStage::Done => return,
-            };
-            self.envelope = EnvelopeStage::Release { t: 0.0, from_level };
-        }
-
-        fn done(&self) -> bool {
-            matches!(self.envelope, EnvelopeStage::Done)
-        }
-
-        fn advance_envelope(&mut self, dt: f32) -> f32 {
-            match &mut self.envelope {
-                EnvelopeStage::Attack { t } => {
-                    *t += dt;
-                    if self.adsr.attack_s <= 0.0 || *t >= self.adsr.attack_s {
-                        self.envelope = EnvelopeStage::Decay { t: 0.0 };
-                        1.0
-                    } else {
-                        *t / self.adsr.attack_s
-                    }
-                }
-                EnvelopeStage::Decay { t } => {
-                    *t += dt;
-                    if self.adsr.decay_s <= 0.0 || *t >= self.adsr.decay_s {
-                        self.envelope = EnvelopeStage::Sustain;
-                        self.adsr.sustain
-                    } else {
-                        (1.0 - self.adsr.sustain).mul_add(-(*t / self.adsr.decay_s), 1.0)
-                    }
-                }
-                EnvelopeStage::Sustain => self.adsr.sustain,
-                EnvelopeStage::Release { t, from_level } => {
-                    *t += dt;
-                    if self.adsr.release_s <= 0.0 || *t >= self.adsr.release_s {
-                        self.envelope = EnvelopeStage::Done;
-                        0.0
-                    } else {
-                        *from_level * (1.0 - (*t / self.adsr.release_s))
-                    }
-                }
-                EnvelopeStage::Done => 0.0,
-            }
-        }
-
-        /// Render the raw waveform at the current phase. Takes `&mut self`
-        /// because the `Noise` wave advances its PRNG and one-pole filter
-        /// state; the periodic waves only read `phase`.
-        fn waveform(&mut self) -> f32 {
-            match self.wave {
-                Wave::Sine => (self.phase * TAU).sin(),
-                Wave::Square => {
-                    if self.phase < 0.5 {
-                        1.0
-                    } else {
-                        -1.0
-                    }
-                }
-                Wave::Triangle => {
-                    if self.phase < 0.5 {
-                        4.0f32.mul_add(self.phase, -1.0)
-                    } else {
-                        4.0f32.mul_add(-self.phase, 3.0)
-                    }
-                }
-                Wave::Saw => 2.0f32.mul_add(-self.phase, 1.0),
-                Wave::Noise { lowpass, tone_mix } => {
-                    let white = next_noise(&mut self.rng);
-                    // One-pole lowpass: y += coeff * (x - y). coeff 1.0
-                    // passes the raw noise; smaller smooths it.
-                    self.lp_prev = lowpass.mul_add(white - self.lp_prev, self.lp_prev);
-                    let noise = self.lp_prev;
-                    if tone_mix > 0.0 {
-                        let tone = (self.phase * TAU).sin();
-                        tone_mix.mul_add(tone, (1.0 - tone_mix) * noise)
-                    } else {
-                        noise
-                    }
-                }
-            }
-        }
-
-        fn next_sample(&mut self, dt: f32) -> f32 {
-            let env = self.advance_envelope(dt);
-            let s = self.waveform() * self.amplitude * env;
-            let ratio = 1.0 + self.sweep_offset;
-            self.phase = self.phase_step.mul_add(ratio, self.phase);
-            if self.phase >= 1.0 {
-                self.phase -= 1.0;
-            }
-            self.sweep_offset *= self.sweep_decay;
-            s
-        }
-    }
-
-    /// One sine partial of a partial-bank voice. `amp` decays toward zero
-    /// each sample by `decay_mul = exp(-rate * dt)`, so the hot loop holds
-    /// no transcendentals beyond the `sin`.
-    #[derive(Copy, Clone, Debug)]
-    struct Partial {
-        phase: f32,
-        phase_step: f32,
-        amp: f32,
-        decay_mul: f32,
-    }
-
-    impl Partial {
-        const SILENT: Self = Self {
-            phase: 0.0,
-            phase_step: 0.0,
-            amp: 0.0,
-            decay_mul: 1.0,
-        };
-    }
-
-    /// Global attack/release ramp wrapping a partial bank. The partials
-    /// carry their own per-sample decay; this ramp swells the voice in at
-    /// `note_on` and damps it out at `note_off`.
-    #[derive(Copy, Clone, Debug)]
-    enum BankStage {
-        Attack { t: f32 },
-        Sustain,
-        Release { t: f32, from_level: f32 },
-        Done,
-    }
-
-    /// The partial-bank voice kernel — a fixed array of inharmonic sine
-    /// partials with per-partial exponential decay, wrapped by a global
-    /// attack/release ramp. Built once at `note_on`; the per-sample path
-    /// is `sin` + multiply-accumulate + decay multiply per partial.
-    #[derive(Copy, Clone, Debug)]
-    struct PartialBankVoice {
-        partials: [Partial; PARTIAL_COUNT],
-        /// Overall level after velocity scaling; the partial amps carry
-        /// the (normalised) spectral shape, this carries the loudness.
-        amplitude: f32,
-        stage: BankStage,
-        attack_s: f32,
-        release_s: f32,
-    }
-
-    impl PartialBankVoice {
-        fn new(
-            pitch: u8,
-            velocity: u8,
-            def: &PartialBankDef,
-            base_amp: f32,
-            sample_rate: f32,
-        ) -> Self {
-            let f0 = 440.0 * ((f32::from(pitch) - 69.0) / 12.0).exp2();
-            let v = f32::from(velocity) / 127.0;
-            let amplitude = base_amp * v;
-            let pitch_scale = f0 / REFERENCE_FREQ;
-            let dt = 1.0 / sample_rate;
-
-            let mut partials = [Partial::SILENT; PARTIAL_COUNT];
-            let mut total = 0.0f32;
-            for (i, p) in partials.iter_mut().enumerate() {
-                // PARTIAL_COUNT is 8 — the index-to-float casts are exact.
-                #[allow(clippy::cast_precision_loss)]
-                let i_f = i as f32;
-                let n = i_f + 1.0;
-                let stretch = (def.inharmonicity * n).mul_add(n, 1.0).sqrt();
-                let detune = if i % 2 == 0 {
-                    1.0 + def.detune
-                } else {
-                    1.0 - def.detune
-                };
-                p.phase_step = (n * f0 * stretch * detune) / sample_rate;
-                let rate = def.decay_base * i_f.mul_add(def.decay_spread, 1.0) * pitch_scale;
-                p.decay_mul = (-rate * dt).exp();
-                let amp = def.partial_amps[i] * i_f.mul_add(def.brightness_tilt * v, 1.0);
-                p.amp = amp;
-                total += amp;
-            }
-            if total > 0.0 {
-                let norm = 1.0 / total;
-                for p in &mut partials {
-                    p.amp *= norm;
-                }
-            }
-
-            Self {
-                partials,
-                amplitude,
-                stage: BankStage::Attack { t: 0.0 },
-                attack_s: def.attack_s,
-                release_s: def.release_s,
-            }
-        }
-
-        fn note_off(&mut self) {
-            let from_level = match self.stage {
-                BankStage::Attack { t } => {
-                    if self.attack_s > 0.0 {
-                        (t / self.attack_s).clamp(0.0, 1.0)
-                    } else {
-                        1.0
-                    }
-                }
-                BankStage::Sustain => 1.0,
-                BankStage::Release { .. } | BankStage::Done => return,
-            };
-            self.stage = BankStage::Release { t: 0.0, from_level };
-        }
-
-        fn done(&self) -> bool {
-            matches!(self.stage, BankStage::Done)
-        }
-
-        fn advance_ramp(&mut self, dt: f32) -> f32 {
-            match &mut self.stage {
-                BankStage::Attack { t } => {
-                    *t += dt;
-                    if self.attack_s <= 0.0 || *t >= self.attack_s {
-                        self.stage = BankStage::Sustain;
-                        1.0
-                    } else {
-                        *t / self.attack_s
-                    }
-                }
-                BankStage::Sustain => 1.0,
-                BankStage::Release { t, from_level } => {
-                    *t += dt;
-                    if self.release_s <= 0.0 || *t >= self.release_s {
-                        self.stage = BankStage::Done;
-                        0.0
-                    } else {
-                        *from_level * (1.0 - (*t / self.release_s))
-                    }
-                }
-                BankStage::Done => 0.0,
-            }
-        }
-
-        fn next_sample(&mut self, dt: f32) -> f32 {
-            let ramp = self.advance_ramp(dt);
-            let mut acc = 0.0f32;
-            let mut amp_sum = 0.0f32;
-            for p in &mut self.partials {
-                acc = (p.phase * TAU).sin().mul_add(p.amp, acc);
-                p.phase += p.phase_step;
-                if p.phase >= 1.0 {
-                    p.phase -= 1.0;
-                }
-                p.amp *= p.decay_mul;
-                amp_sum += p.amp.abs();
-            }
-            // A held voice whose partials have rung out frees itself; a
-            // pad (zero partial decay) only ends once its release ramp
-            // completes.
-            if matches!(self.stage, BankStage::Sustain)
-                && amp_sum * self.amplitude < PARTIAL_SILENCE_FLOOR
-            {
-                self.stage = BankStage::Done;
-            }
-            acc * self.amplitude * ramp
-        }
-
-        #[cfg(test)]
-        fn envelope_level(&self) -> f32 {
-            self.partials.iter().map(|p| p.amp.abs()).sum()
-        }
-
-        #[cfg(test)]
-        fn partial_amps(&self) -> [f32; PARTIAL_COUNT] {
-            let mut out = [0.0f32; PARTIAL_COUNT];
-            for (slot, p) in out.iter_mut().zip(self.partials.iter()) {
-                *slot = p.amp;
-            }
-            out
-        }
-
-        #[cfg(test)]
-        fn in_sustain(&self) -> bool {
-            matches!(self.stage, BankStage::Sustain)
-        }
-    }
-
-    /// Attack ramp (seconds) wrapping a sample voice — a short swell so a
-    /// re-pitched recording doesn't click on at full level (ADR-0103 §6,
-    /// the partial bank's ramp shape).
-    const SAMPLE_ATTACK_SECS: f32 = 0.003;
-    /// Release ramp (seconds) on `note_off` for a sample voice — the
-    /// damper that ends a held note faster than the sample's natural decay.
-    const SAMPLE_RELEASE_SECS: f32 = 0.08;
-    /// Base amplitude of a sample voice before velocity scaling. Sampled
-    /// recordings already carry their own level; this trims headroom so a
-    /// dense chord doesn't clip past the soft-clip.
-    const SAMPLE_BASE_AMP: f32 = 0.6;
-
-    /// A region's sustain loop in device-rate coordinates (ADR-0103 §6).
-    /// The SFZ frame offsets are scaled at bank assembly by the load-time
-    /// resample ratio into these fractional positions, so the kernel wrap
-    /// interpolates sub-sample and rounding never lands a click. `start`
-    /// and `end` index the region's device-rate PCM; the voice cycles the
-    /// half-open `[start, end)` interval while it sounds.
-    #[derive(Clone, Copy, Debug, PartialEq)]
-    struct SampleLoop {
-        start: f32,
-        end: f32,
-    }
-
-    /// One region of a sampled instrument bank (ADR-0103 §5/§6): a
-    /// device-rate mono recording plus the inclusive MIDI key range it
-    /// covers, the inclusive velocity range it answers to, the root pitch
-    /// it was recorded at (so a voice repitches by
-    /// `2^((pitch − pitch_keycenter) / 12)`), and an optional sustain loop.
-    /// The PCM is `Arc`'d so every region naming the same sample shares one
-    /// buffer and a spawned voice holds a cheap reference, not a copy.
-    #[derive(Clone, Debug)]
-    struct SampleRegion {
-        lokey: u8,
-        hikey: u8,
-        lovel: u8,
-        hivel: u8,
-        pitch_keycenter: u8,
-        pcm: Arc<[f32]>,
-        /// The sustain loop, or `None` for a full-decay region that plays
-        /// once and ends when its sample runs out.
-        loop_region: Option<SampleLoop>,
-    }
-
-    /// A loaded sampled-instrument bank (ADR-0103 §4/§5): the regions to
-    /// select between by `(pitch, velocity)`, the name derived from the
-    /// `.sfz` filename, and the total decoded PCM the bank holds resident
-    /// (reported in the load reply — there is no unload in v1).
-    #[derive(Debug)]
-    struct SampleBank {
-        name: String,
-        regions: Vec<SampleRegion>,
-        resident_bytes: usize,
-    }
-
-    impl SampleBank {
-        /// The first region whose key and velocity ranges both contain
-        /// `(pitch, velocity)`, or `None` when the note falls in a gap the
-        /// bank doesn't cover (the `note_on` then drops).
-        fn select(&self, pitch: u8, velocity: u8) -> Option<&SampleRegion> {
-            self.regions.iter().find(|r| {
-                (r.lokey..=r.hikey).contains(&pitch) && (r.lovel..=r.hivel).contains(&velocity)
-            })
-        }
-    }
-
-    /// The sample voice kernel (ADR-0103 §6): walk the region's device-rate
-    /// PCM at a repitched rate with linear interpolation, wrapped in the
-    /// same short attack / `note_off`-release ramp the partial bank uses.
-    /// An unlooped region ends when its sample runs out (full-decay,
-    /// piano-class sets). A looped region cycles `[loop_start, loop_end)`
-    /// while it sounds — interpolating across the seam back to
-    /// `loop_start` — and holds the note indefinitely, ending only once the
-    /// `note_off` release ramp completes (the loop keeps cycling beneath
-    /// the fade).
-    #[derive(Clone, Debug)]
-    struct SampleVoice {
-        /// Device-rate mono PCM of the selected region, shared with the
-        /// bank.
-        pcm: Arc<[f32]>,
-        /// Fractional read position into `pcm`, advanced by `rate` each
-        /// output sample.
-        pos: f32,
-        /// Playback rate ratio `2^((pitch − pitch_keycenter) / 12)` — the
-        /// repitch from the region's root note. The PCM is already at the
-        /// device rate, so this is the only resampling the hot loop does.
-        rate: f32,
-        /// Velocity-scaled amplitude the interpolated sample is multiplied
-        /// by.
-        amplitude: f32,
-        /// The sustain loop bounds (device-rate fractional positions), or
-        /// `None` for an unlooped region.
-        loop_region: Option<SampleLoop>,
-        /// Attack / release ramp, the partial bank's shape.
-        stage: BankStage,
-        attack_s: f32,
-        release_s: f32,
-        /// Set once an unlooped region's read position walks off the end of
-        /// the PCM, or the release ramp completes. A looped voice never sets
-        /// it from exhaustion — it ends through the release ramp.
-        finished: bool,
-    }
-
-    impl SampleVoice {
-        fn new(pitch: u8, velocity: u8, region: &SampleRegion) -> Self {
-            let semitones = f32::from(pitch) - f32::from(region.pitch_keycenter);
-            let rate = (semitones / 12.0).exp2();
-            let v = f32::from(velocity) / 127.0;
-            // Drop a loop whose bounds collapsed (defensive — assembly only
-            // emits `start < end`): a non-positive span has no cycle.
-            let loop_region = region
-                .loop_region
-                .filter(|lp| lp.end > lp.start && lp.start >= 0.0);
-            Self {
-                pcm: Arc::clone(&region.pcm),
-                pos: 0.0,
-                rate,
-                amplitude: SAMPLE_BASE_AMP * v,
-                loop_region,
-                stage: BankStage::Attack { t: 0.0 },
-                attack_s: SAMPLE_ATTACK_SECS,
-                release_s: SAMPLE_RELEASE_SECS,
-                finished: false,
-            }
-        }
-
-        fn note_off(&mut self) {
-            let from_level = match self.stage {
-                BankStage::Attack { t } => {
-                    if self.attack_s > 0.0 {
-                        (t / self.attack_s).clamp(0.0, 1.0)
-                    } else {
-                        1.0
-                    }
-                }
-                BankStage::Sustain => 1.0,
-                BankStage::Release { .. } | BankStage::Done => return,
-            };
-            self.stage = BankStage::Release { t: 0.0, from_level };
-        }
-
-        fn done(&self) -> bool {
-            self.finished || matches!(self.stage, BankStage::Done)
-        }
-
-        /// Advance the attack/release ramp one sample, returning its current
-        /// level — the partial bank's ramp logic over the sample voice's
-        /// own attack/release times.
-        fn advance_ramp(&mut self, dt: f32) -> f32 {
-            match &mut self.stage {
-                BankStage::Attack { t } => {
-                    *t += dt;
-                    if self.attack_s <= 0.0 || *t >= self.attack_s {
-                        self.stage = BankStage::Sustain;
-                        1.0
-                    } else {
-                        *t / self.attack_s
-                    }
-                }
-                BankStage::Sustain => 1.0,
-                BankStage::Release { t, from_level } => {
-                    *t += dt;
-                    if self.release_s <= 0.0 || *t >= self.release_s {
-                        self.stage = BankStage::Done;
-                        0.0
-                    } else {
-                        *from_level * (1.0 - (*t / self.release_s))
-                    }
-                }
-                BankStage::Done => 0.0,
-            }
-        }
-
-        // Read position and PCM lengths are bounded well below 2^24 for any
-        // sane sample, so the index-to-float and float-to-index casts in the
-        // looped / unlooped readers are exact and non-negative on the hot
-        // path.
-        fn next_sample(&mut self, dt: f32) -> f32 {
-            let ramp = self.advance_ramp(dt);
-            if self.finished {
-                return 0.0;
-            }
-            let len = self.pcm.len();
-            if len == 0 {
-                self.finished = true;
-                return 0.0;
-            }
-            match self.loop_region {
-                Some(lp) => self.next_looped(lp, len, ramp),
-                None => self.next_unlooped(len, ramp),
-            }
-        }
-
-        /// The unlooped read: linear interpolation over the PCM, ending the
-        /// voice once the read position walks off the end (ADR-0103 §6).
-        #[allow(
-            clippy::cast_precision_loss,
-            clippy::cast_possible_truncation,
-            clippy::cast_sign_loss
-        )]
-        fn next_unlooped(&mut self, len: usize, ramp: f32) -> f32 {
-            let i = self.pos.floor() as usize;
-            if i >= len {
-                self.finished = true;
-                return 0.0;
-            }
-            let a = self.pcm[i];
-            let b = self.pcm[(i + 1).min(len - 1)];
-            let frac = self.pos - i as f32;
-            let s = (b - a).mul_add(frac, a) * self.amplitude * ramp;
-            self.pos += self.rate;
-            if self.pos >= len as f32 {
-                // The held note's sample ran out — an unlooped voice ends.
-                self.finished = true;
-            }
-            s
-        }
-
-        /// The looped read (ADR-0103 §6): interpolate within `[loop_start,
-        /// loop_end)`, wrapping back to `loop_start` at the seam so the
-        /// interpolation reads `loop_start` as the post-seam neighbour and
-        /// produces no discontinuity beyond interpolation error. The voice
-        /// never ends from exhaustion here — only the release ramp retires
-        /// it (the loop keeps cycling beneath the fade).
-        #[allow(
-            clippy::cast_precision_loss,
-            clippy::cast_possible_truncation,
-            clippy::cast_sign_loss
-        )]
-        fn next_looped(&mut self, lp: SampleLoop, len: usize, ramp: f32) -> f32 {
-            // `pos < loop_end <= len` holds going in, so `i` is in range.
-            let i = (self.pos.floor() as usize).min(len - 1);
-            let a = self.pcm[i];
-            // The interpolation neighbour is the next frame — but if that
-            // frame reaches or crosses `loop_end`, the loop wraps, so read
-            // `loop_start` instead (the seam neighbour).
-            let next_index = if (i + 1) as f32 >= lp.end {
-                (lp.start.floor() as usize).min(len - 1)
-            } else {
-                i + 1
-            };
-            let b = self.pcm[next_index];
-            let frac = self.pos - i as f32;
-            let s = (b - a).mul_add(frac, a) * self.amplitude * ramp;
-
-            self.pos += self.rate;
-            if self.pos >= lp.end {
-                // Wrap the overshoot back into the loop region. Modulo the
-                // loop length so a rate larger than the span still lands in
-                // `[loop_start, loop_end)` in O(1).
-                let loop_len = lp.end - lp.start;
-                let over = self.pos - lp.end;
-                let wrapped = (over / loop_len).floor().mul_add(-loop_len, over);
-                self.pos = lp.start + wrapped;
-            }
-            s
-        }
-    }
-
-    /// Voice kernel — one of the three synthesis models, selected by the
-    /// instrument at `note_on`: a built-in oscillator or partial-bank
-    /// patch, or a loaded sampled instrument (ADR-0103 §6).
-    #[derive(Clone, Debug)]
-    enum VoiceKernel {
-        Oscillator(OscVoice),
-        PartialBank(PartialBankVoice),
-        Sample(SampleVoice),
-    }
-
-    /// Build the kernel for a built-in instrument patch (oscillator or
-    /// partial bank). Split out of [`Voice`] so the `note_on` path can
-    /// resolve a built-in or a loaded sample bank into a `VoiceKernel`
-    /// before the steal / dedup bookkeeping, then stamp one `Voice`.
-    fn build_builtin_kernel(
-        sender_mailbox: MailboxId,
-        instrument_id: u8,
-        pitch: u8,
-        velocity: u8,
-        def: &InstrumentDef,
-        sample_rate: f32,
-    ) -> VoiceKernel {
-        match def.voice {
-            VoiceDef::Oscillator { wave, adsr } => {
-                let seed = voice_seed(sender_mailbox, instrument_id, pitch);
-                let mut osc =
-                    OscVoice::new(pitch, velocity, wave, adsr, def.base_amp, sample_rate, seed);
-                if let Some(sweep) = def.pitch_sweep {
-                    osc = osc.with_pitch_sweep(sweep, sample_rate);
-                }
-                VoiceKernel::Oscillator(osc)
-            }
-            VoiceDef::PartialBank(bank) => VoiceKernel::PartialBank(PartialBankVoice::new(
-                pitch,
-                velocity,
-                &bank,
-                def.base_amp,
-                sample_rate,
-            )),
-        }
-    }
-
-    /// A single sounding voice: the routing key (`sender_mailbox`,
-    /// `instrument_id`, `pitch`) plus the kernel that renders it. No longer
-    /// `Copy` — a sample voice holds a reference-counted PCM handle
-    /// (ADR-0103 §6) — but the pool was never structurally dependent on
-    /// `Copy`; it stays a flat `Vec<Voice>` mutated by `swap_remove` /
-    /// `push`.
-    ///
-    /// `seq` is a monotonically increasing counter stamped at allocation,
-    /// used by voice-steal to locate the oldest voice regardless of the
-    /// pool's current order (which `swap_remove` scrambles).
-    #[derive(Clone, Debug)]
-    struct Voice {
-        sender_mailbox: MailboxId,
-        instrument_id: u8,
-        pitch: u8,
-        seq: u64,
-        kernel: VoiceKernel,
-    }
-
-    impl Voice {
-        fn note_off(&mut self) {
-            match &mut self.kernel {
-                VoiceKernel::Oscillator(v) => v.note_off(),
-                VoiceKernel::PartialBank(v) => v.note_off(),
-                VoiceKernel::Sample(v) => v.note_off(),
-            }
-        }
-
-        fn done(&self) -> bool {
-            match &self.kernel {
-                VoiceKernel::Oscillator(v) => v.done(),
-                VoiceKernel::PartialBank(v) => v.done(),
-                VoiceKernel::Sample(v) => v.done(),
-            }
-        }
-
-        fn next_sample(&mut self, dt: f32) -> f32 {
-            match &mut self.kernel {
-                VoiceKernel::Oscillator(v) => v.next_sample(dt),
-                VoiceKernel::PartialBank(v) => v.next_sample(dt),
-                VoiceKernel::Sample(v) => v.next_sample(dt),
-            }
-        }
-    }
-
-    /// Fade state of a [`TrackVoice`]. A track plays at full level until
-    /// `stop_track` arms a short linear fade-out; `remaining` counts down
-    /// per output sample and the track retires when it hits zero (ADR-0103
-    /// §3).
-    #[derive(Clone, Debug)]
-    enum TrackFade {
-        Playing,
-        FadingOut { remaining: u32, total: u32 },
-    }
-
-    /// One playing track in the dedicated mixer lane (ADR-0103 §3). Holds
-    /// the `Arc`'d device-rate mono PCM, a position walk, per-track gain,
-    /// loop flag, and fade state. A track neither counts against
-    /// `MAX_VOICES` nor participates in voice-steal — a music bed must not
-    /// be evicted by a note flurry. Keyed by `(sender_mailbox, lane,
-    /// namespace, path)`, mirroring the voice key plus the caller-supplied
-    /// `lane` that disambiguates senders sharing a source mailbox.
-    struct TrackVoice {
-        sender_mailbox: MailboxId,
-        lane: Option<String>,
-        namespace: String,
-        path: String,
-        pcm: Arc<[f32]>,
-        position: usize,
-        gain: f32,
-        looping: bool,
-        fade: TrackFade,
-        done: bool,
-    }
-
-    impl TrackVoice {
-        fn new(
-            sender_mailbox: MailboxId,
-            lane: Option<String>,
-            namespace: String,
-            path: String,
-            pcm: Arc<[f32]>,
-            gain: f32,
-            looping: bool,
-        ) -> Self {
-            Self {
-                sender_mailbox,
-                lane,
-                namespace,
-                path,
-                pcm,
-                position: 0,
-                gain,
-                looping,
-                fade: TrackFade::Playing,
-                done: false,
-            }
-        }
-
-        /// True when this event's key matches the track's
-        /// `(sender_mailbox, lane, namespace, path)`.
-        fn matches(
-            &self,
-            sender_mailbox: MailboxId,
-            lane: Option<&String>,
-            namespace: &str,
-            path: &str,
-        ) -> bool {
-            self.sender_mailbox == sender_mailbox
-                && self.lane.as_ref() == lane
-                && self.namespace == namespace
-                && self.path == path
-        }
-
-        /// Arm the fade-out. Idempotent — a second `stop` while already
-        /// fading keeps the first fade's progress.
-        fn stop(&mut self, fade_samples: u32) {
-            if matches!(self.fade, TrackFade::Playing) {
-                let total = fade_samples.max(1);
-                self.fade = TrackFade::FadingOut {
-                    remaining: total,
-                    total,
-                };
-            }
-        }
-
-        fn done(&self) -> bool {
-            self.done
-        }
-
-        /// Render this track's next sample (already gained + faded) and
-        /// advance the position. Returns `0.0` once retired; an empty PCM
-        /// buffer retires immediately.
-        fn next_sample(&mut self) -> f32 {
-            if self.done || self.pcm.is_empty() {
-                self.done = true;
-                return 0.0;
-            }
-            let fade_mul = match &mut self.fade {
-                TrackFade::Playing => 1.0,
-                TrackFade::FadingOut { remaining, total } => {
-                    if *remaining == 0 {
-                        self.done = true;
-                        return 0.0;
-                    }
-                    // `remaining` / `total` are small fade-window counts —
-                    // the ratio is exact in f32.
-                    #[allow(clippy::cast_precision_loss)]
-                    let mul = *remaining as f32 / *total as f32;
-                    *remaining -= 1;
-                    mul
-                }
-            };
-            let sample = self.pcm[self.position] * self.gain * fade_mul;
-            self.position += 1;
-            if self.position >= self.pcm.len() {
-                if self.looping {
-                    self.position = 0;
-                } else {
-                    self.done = true;
-                }
-            }
-            sample
-        }
-    }
-
-    /// Whole-process synth state. Lives on the cpal callback thread;
-    /// the cap communicates via the event queue.
-    struct Synth {
-        events: Arc<ArrayQueue<AudioEvent>>,
-        voices: Vec<Voice>,
-        /// Track playback lane (ADR-0103 §3) — separate from `voices` so a
-        /// track is never counted against `MAX_VOICES` nor voice-stolen.
-        tracks: Vec<TrackVoice>,
-        /// Loaded sampled-instrument banks (ADR-0103 §4), appended in load
-        /// order. Index `i` is `instrument_id` `BUILTINS.len() + i`, so a
-        /// `note_on` whose id walks past the built-ins indexes here. The cap
-        /// assigns ids the same way, so the two stay in lockstep.
-        banks: Vec<Arc<SampleBank>>,
-        sample_rate: f32,
-        master_gain: f32,
-        /// Monotonically increasing counter stamped into each `Voice::seq`
-        /// at allocation. Voice-steal uses the minimum value to locate the
-        /// oldest voice regardless of pool order.
-        next_seq: u64,
-        /// Running output-frame counter (ADR-0104). Advanced by the frame
-        /// count of every `fill`; the timebase scheduled events are placed
-        /// against and fire from. Callback-owned, so no locking.
-        frame_clock: u64,
-        /// Pending scheduled note events ordered by due frame (ADR-0104),
-        /// a min-heap via `Reverse`. `fill` pops the events that fall on
-        /// each frame and routes them through the note-on / note-off paths.
-        scheduled: BinaryHeap<Reverse<ScheduledEntry>>,
-        /// Monotonic stamp threaded into each `ScheduledEntry::seq` so that
-        /// events on the same due frame fire in batch-arrival order.
-        next_schedule_seq: u64,
-    }
-
-    impl Synth {
-        fn new(events: Arc<ArrayQueue<AudioEvent>>, sample_rate: f32) -> Self {
-            Self {
-                events,
-                voices: Vec::with_capacity(MAX_VOICES),
-                tracks: Vec::new(),
-                banks: Vec::new(),
-                sample_rate,
-                master_gain: 1.0,
-                next_seq: 0,
-                frame_clock: 0,
-                scheduled: BinaryHeap::new(),
-                next_schedule_seq: 0,
-            }
-        }
-
-        /// Resolve a loaded sample bank by `instrument_id`, returning a
-        /// cheap `Arc` clone (or `None` for an id still inside the built-in
-        /// range or past the loaded banks). The `note_on` path falls back
-        /// to this when `instrument_by_id` misses.
-        fn bank_for(&self, instrument_id: u8) -> Option<Arc<SampleBank>> {
-            let index = (instrument_id as usize).checked_sub(BUILTINS.len())?;
-            self.banks.get(index).map(Arc::clone)
-        }
-
-        /// Number of output samples in the `stop_track` fade-out at this
-        /// device rate.
-        fn fade_samples(&self) -> u32 {
-            // Fade window is a few milliseconds at audio rates — well
-            // within u32 and non-negative.
-            #[allow(clippy::cast_possible_truncation, clippy::cast_sign_loss)]
-            let n = (TRACK_FADE_SECS * self.sample_rate) as u32;
-            n
-        }
-
-        /// Admit a `note_on`: resolve its kernel (a built-in patch, or — when
-        /// the id walks past the built-ins — a loaded sample bank's region
-        /// selected by `(pitch, velocity)`), then steal the oldest voice if
-        /// at capacity, replace any voice already on the same key, and push.
-        /// A miss on both kernel sources (unknown id, or a bank with no
-        /// region covering the note) warn-drops without touching the pool
-        /// (ADR-0103 §6).
-        fn trigger_note_on(
-            &mut self,
-            sender_mailbox: MailboxId,
-            pitch: u8,
-            velocity: u8,
-            instrument_id: u8,
-        ) {
-            let kernel = if let Some(def) = instrument_by_id(instrument_id) {
-                Some(build_builtin_kernel(
-                    sender_mailbox,
-                    instrument_id,
-                    pitch,
-                    velocity,
-                    def,
-                    self.sample_rate,
-                ))
-            } else {
-                self.bank_for(instrument_id).and_then(|bank| {
-                    bank.select(pitch, velocity).map(|region| {
-                        VoiceKernel::Sample(SampleVoice::new(pitch, velocity, region))
-                    })
-                })
-            };
-            let Some(kernel) = kernel else {
-                tracing::warn!(
-                    target: "aether_substrate::audio",
-                    instrument_id,
-                    pitch,
-                    velocity,
-                    "note_on: no instrument / region for id, dropping",
-                );
-                return;
-            };
-            if self.voices.len() >= MAX_VOICES {
-                // Evict the oldest (minimum-seq) voice. swap_remove is O(1)
-                // and safe here because the pool is non-empty at capacity.
-                if let Some(oldest_idx) = self
-                    .voices
-                    .iter()
-                    .enumerate()
-                    .min_by_key(|(_, v)| v.seq)
-                    .map(|(i, _)| i)
-                {
-                    self.voices.swap_remove(oldest_idx);
-                }
-            }
-            if let Some(existing) = self.voices.iter().position(|v| {
-                v.sender_mailbox == sender_mailbox
-                    && v.instrument_id == instrument_id
-                    && v.pitch == pitch
-            }) {
-                self.voices.swap_remove(existing);
-            }
-            let seq = self.next_seq;
-            self.next_seq += 1;
-            self.voices.push(Voice {
-                sender_mailbox,
-                instrument_id,
-                pitch,
-                seq,
-                kernel,
-            });
-        }
-
-        /// Release the voice matching `(sender_mailbox, instrument_id,
-        /// pitch)`, if one is sounding. A miss is a silent no-op (a late or
-        /// unmatched note-off), matching the immediate `note_off` path.
-        /// Shared by the queue-drained note-off and the scheduled note-off.
-        fn trigger_note_off(&mut self, sender_mailbox: MailboxId, pitch: u8, instrument_id: u8) {
-            if let Some(v) = self.voices.iter_mut().find(|v| {
-                v.sender_mailbox == sender_mailbox
-                    && v.instrument_id == instrument_id
-                    && v.pitch == pitch
-            }) {
-                v.note_off();
-            }
-        }
-
-        /// Fire one scheduled note event through the same paths the
-        /// immediate mail would take (ADR-0104).
-        fn fire_scheduled(&mut self, sender_mailbox: MailboxId, note: &ScheduledNote) {
-            match *note {
-                ScheduledNote::On {
-                    pitch,
-                    velocity,
-                    instrument_id,
-                } => self.trigger_note_on(sender_mailbox, pitch, velocity, instrument_id),
-                ScheduledNote::Off {
-                    pitch,
-                    instrument_id,
-                } => self.trigger_note_off(sender_mailbox, pitch, instrument_id),
-            }
-        }
-
-        fn drain_events(&mut self) {
-            while let Some(ev) = self.events.pop() {
-                match ev {
-                    AudioEvent::NoteOn {
-                        sender_mailbox,
-                        pitch,
-                        velocity,
-                        instrument_id,
-                    } => self.trigger_note_on(sender_mailbox, pitch, velocity, instrument_id),
-                    AudioEvent::NoteOff {
-                        sender_mailbox,
-                        pitch,
-                        instrument_id,
-                    } => self.trigger_note_off(sender_mailbox, pitch, instrument_id),
-                    AudioEvent::SetMasterGain { gain } => {
-                        self.master_gain = gain.clamp(0.0, 1.0);
-                    }
-                    AudioEvent::TrackStart {
-                        sender_mailbox,
-                        lane,
-                        namespace,
-                        path,
-                        pcm,
-                        gain,
-                        looping,
-                    } => {
-                        self.start_track(sender_mailbox, lane, namespace, path, pcm, gain, looping);
-                    }
-                    AudioEvent::TrackStop {
-                        sender_mailbox,
-                        lane,
-                        namespace,
-                        path,
-                    } => self.stop_track(sender_mailbox, lane.as_ref(), &namespace, &path),
-                    AudioEvent::RegisterInstrument { id, bank } => {
-                        // Banks arrive in load order on this single-producer
-                        // FIFO, and the cap assigns ids from `BUILTINS.len()`
-                        // upward in the same order, so the new bank's index
-                        // is exactly `id - BUILTINS.len()` == current length.
-                        // A mismatch is a wiring bug, not a runtime input —
-                        // log it but still append so lookups stay dense.
-                        let expected = BUILTINS.len() + self.banks.len();
-                        if id as usize != expected {
-                            tracing::warn!(
-                                target: "aether_substrate::audio",
-                                id,
-                                expected,
-                                "register_instrument: id out of step with load order",
-                            );
-                        }
-                        self.banks.push(bank);
-                    }
-                    AudioEvent::Schedule {
-                        sender_mailbox,
-                        events,
-                    } => {
-                        // Offsets are relative to receipt at the callback —
-                        // the current frame clock (this drain runs at block
-                        // start). Every event in the batch shares this
-                        // anchor, so simultaneous events stay simultaneous.
-                        for event in events {
-                            let due_frame = self.frame_clock
-                                + millis_to_frames(event.at_millis, self.sample_rate);
-                            let seq = self.next_schedule_seq;
-                            self.next_schedule_seq += 1;
-                            self.scheduled.push(Reverse(ScheduledEntry {
-                                due_frame,
-                                seq,
-                                sender_mailbox,
-                                note: event.event,
-                            }));
-                        }
-                    }
-                }
-            }
-        }
-
-        /// Start (or restart) a track in the lane. Re-playing the same
-        /// `(sender_mailbox, lane, namespace, path)` key drops the existing
-        /// track first, so a key never stacks.
-        #[allow(clippy::too_many_arguments)]
-        fn start_track(
-            &mut self,
-            sender_mailbox: MailboxId,
-            lane: Option<String>,
-            namespace: String,
-            path: String,
-            pcm: Arc<[f32]>,
-            gain: f32,
-            looping: bool,
-        ) {
-            if let Some(i) = self
-                .tracks
-                .iter()
-                .position(|t| t.matches(sender_mailbox, lane.as_ref(), &namespace, &path))
-            {
-                self.tracks.swap_remove(i);
-            }
-            self.tracks.push(TrackVoice::new(
-                sender_mailbox,
-                lane,
-                namespace,
-                path,
-                pcm,
-                gain,
-                looping,
-            ));
-        }
-
-        /// Arm the fade-out on the track at this key, if one is playing.
-        fn stop_track(
-            &mut self,
-            sender_mailbox: MailboxId,
-            lane: Option<&String>,
-            namespace: &str,
-            path: &str,
-        ) {
-            let fade = self.fade_samples();
-            if let Some(t) = self
-                .tracks
-                .iter_mut()
-                .find(|t| t.matches(sender_mailbox, lane, namespace, path))
-            {
-                t.stop(fade);
-            }
-        }
-
-        fn fill(&mut self, buffer: &mut [f32], channels: usize) {
-            self.drain_events();
-            let dt = 1.0 / self.sample_rate;
-            let frames = buffer.len() / channels.max(1);
-            for frame in 0..frames {
-                // Fire every scheduled event due on or before this frame
-                // before rendering it, so a scheduled note's voice is alive
-                // for the sample it falls on — sample-accurate by
-                // construction (ADR-0104).
-                let absolute = self.frame_clock + frame as u64;
-                loop {
-                    match self.scheduled.peek() {
-                        Some(Reverse(top)) if top.due_frame <= absolute => {}
-                        _ => break,
-                    }
-                    let Reverse(entry) = self
-                        .scheduled
-                        .pop()
-                        .expect("peeked entry is present this iteration");
-                    self.fire_scheduled(entry.sender_mailbox, &entry.note);
-                }
-                let mut sample = 0.0f32;
-                for voice in &mut self.voices {
-                    sample += voice.next_sample(dt);
-                }
-                // Tracks mix in their own lane, summed after the voices
-                // and before master gain + the soft clip (ADR-0103 §3).
-                for track in &mut self.tracks {
-                    sample += track.next_sample();
-                }
-                sample *= self.master_gain;
-                sample = sample.tanh();
-                let start = frame * channels;
-                for ch in 0..channels {
-                    buffer[start + ch] = sample;
-                }
-            }
-            // Advance the clock by this block so the next drain anchors
-            // scheduled offsets against the right receipt frame (ADR-0104).
-            self.frame_clock += frames as u64;
-            let mut i = 0;
-            while i < self.voices.len() {
-                if self.voices[i].done() {
-                    self.voices.swap_remove(i);
-                } else {
-                    i += 1;
-                }
-            }
-            let mut t = 0;
-            while t < self.tracks.len() {
-                if self.tracks[t].done() {
-                    self.tracks.swap_remove(t);
-                } else {
-                    t += 1;
-                }
-            }
-        }
-
-        #[cfg(test)]
-        fn voice_count(&self) -> usize {
-            self.voices.len()
-        }
-
-        #[cfg(test)]
-        fn has_voice_with_pitch(&self, pitch: u8) -> bool {
-            self.voices.iter().any(|v| v.pitch == pitch)
-        }
-
-        #[cfg(test)]
-        fn master_gain_value(&self) -> f32 {
-            self.master_gain
-        }
-
-        #[cfg(test)]
-        fn track_count(&self) -> usize {
-            self.tracks.len()
-        }
-
-        #[cfg(test)]
-        fn bank_count(&self) -> usize {
-            self.banks.len()
-        }
-
-        #[cfg(test)]
-        fn scheduled_count(&self) -> usize {
-            self.scheduled.len()
-        }
-    }
-
-    /// Handle to a running cpal pipeline. Lives on the audio worker
-    /// thread for the entire run — `cpal::Stream` is `!Send` on macOS,
-    /// so the stream is constructed on, owned by, and dropped from the
-    /// same thread. Dropping the pipeline silences every voice and tears
-    /// down the cpal stream.
-    struct AudioPipeline {
-        sender: AudioEventSender,
-        /// The device output rate the synth runs at. The cap reads it back
-        /// (via the init channel) as the resample target for track decode
-        /// (ADR-0103 §1) — decode happens on the dispatcher, not here.
-        sample_rate: u32,
-        _stream: cpal::Stream,
-    }
-
-    #[derive(Debug)]
-    enum AudioBuildError {
-        NoDevice,
-        RateUnsupported(u32),
-        ConfigQuery(String),
-        StreamBuild(String),
-        StreamPlay(String),
-    }
-
-    impl fmt::Display for AudioBuildError {
-        fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-            match self {
-                Self::NoDevice => write!(f, "no default audio output device"),
-                Self::RateUnsupported(r) => write!(f, "requested sample rate {r} Hz unsupported"),
-                Self::ConfigQuery(e) => write!(f, "config query failed: {e}"),
-                Self::StreamBuild(e) => write!(f, "stream build failed: {e}"),
-                Self::StreamPlay(e) => write!(f, "stream play failed: {e}"),
-            }
-        }
-    }
-
-    fn try_build_pipeline(
-        requested_sample_rate: Option<u32>,
-    ) -> Result<AudioPipeline, AudioBuildError> {
-        let host = cpal::default_host();
-        let device = host
-            .default_output_device()
-            .ok_or(AudioBuildError::NoDevice)?;
-
-        let config = match requested_sample_rate {
-            Some(rate) => {
-                find_config_for_rate(&device, rate).ok_or(AudioBuildError::RateUnsupported(rate))?
-            }
-            None => device
-                .default_output_config()
-                .map_err(|e| AudioBuildError::ConfigQuery(e.to_string()))?
-                .config(),
-        };
-
-        let sample_rate = config.sample_rate;
-        let channels = config.channels;
-
-        let (sender, queue) = new_event_channel();
-        // Audio sample rates are bounded well below 2^24 — exact in f32.
-        #[allow(clippy::cast_precision_loss)]
-        let mut synth = Synth::new(queue, sample_rate as f32);
-
-        let stream = device
-            .build_output_stream(
-                &config,
-                move |data: &mut [f32], _: &cpal::OutputCallbackInfo| {
-                    synth.fill(data, channels as usize);
-                },
-                |err| {
-                    tracing::warn!(
-                        target: "aether_substrate::audio",
-                        error = %err,
-                        "cpal stream error",
-                    );
-                },
-                None,
-            )
-            .map_err(|e| AudioBuildError::StreamBuild(e.to_string()))?;
-
-        stream
-            .play()
-            .map_err(|e| AudioBuildError::StreamPlay(e.to_string()))?;
-
-        tracing::info!(
-            target: "aether_substrate::audio",
-            sample_rate,
-            channels,
-            instruments = builtin_count(),
-            builtin_names = ?builtin_names(),
-            "audio pipeline started",
-        );
-
-        Ok(AudioPipeline {
-            sender,
-            sample_rate,
-            _stream: stream,
-        })
-    }
-
-    fn find_config_for_rate(device: &cpal::Device, rate: u32) -> Option<cpal::StreamConfig> {
-        let configs = device.supported_output_configs().ok()?;
-        for cfg in configs {
-            let min = cfg.min_sample_rate();
-            let max = cfg.max_sample_rate();
-            if rate >= min && rate <= max {
-                return Some(cfg.with_sample_rate(rate).config());
-            }
-        }
-        None
-    }
+    use super::config::AudioConfig;
+    use super::event::{AudioEvent, AudioEventSender};
+    use super::instrument::builtin_id_ceiling;
+    use super::sample::{
+        BankAssembly, BankAssemblyContext, BankAssemblyOutput, PendingInstrument, SampleSlot,
+        assemble_bank, bank_name_from_path, join_fs, sfz_dir,
+    };
+    use super::schedule::{SCHEDULE_MAX_EVENTS, SCHEDULE_MAX_MILLIS};
+    use super::synth::{AudioBuildError, try_build_pipeline};
+    use super::track::{DecodeOutput, PendingTrack, TrackDecodeContext};
+
+    // Items the inline `mod tests` reaches through `use super::*` but the
+    // non-test cap code doesn't — gated so the lib build doesn't see them
+    // as unused. The synth / voice / sample internals the unit tests drive
+    // directly, plus the `AudioConfigLayer` the config-resolution tests
+    // build and the bare kinds / sfz types the scheduling + bank tests use.
+    #[cfg(test)]
+    use super::config::AudioConfigLayer;
+    #[cfg(test)]
+    use super::event::new_event_channel;
+    #[cfg(test)]
+    use super::instrument::{
+        Adsr, BUILTINS, PARTIAL_COUNT, PartialBankDef, PitchSweep, VoiceDef, Wave, builtin_count,
+        builtin_names,
+    };
+    #[cfg(test)]
+    use super::kinds::{ScheduledEvent, ScheduledNote};
+    #[cfg(test)]
+    use super::sample::{SampleBank, SampleLoop, SampleRegion, SampleVoice};
+    #[cfg(test)]
+    use super::sfz::{SfzLoop, SfzRegion};
+    #[cfg(test)]
+    use super::synth::Synth;
+    #[cfg(test)]
+    use super::voice::{MAX_VOICES, OscVoice, PartialBankVoice, voice_seed};
+    #[cfg(test)]
+    use crossbeam_queue::ArrayQueue;
+    #[cfg(test)]
+    use std::f32::consts::TAU;
 
     /// Extract the sender's mailbox id for voice-table keying. Component
     /// senders come through as `EngineMailbox { mailbox_id }`; Claude
@@ -2054,120 +198,6 @@ mod native {
             _ => MailboxId(0),
         }
     }
-
-    /// `aether.audio` mailbox cap. Holds the producer side of the synth
-    /// event queue (the crate-internal `AudioEventSender`), the audio
-    /// worker thread that owns the [`cpal::Stream`] (see module-level
-    /// "per-cap audio worker" docs for the `!Send` rationale), and a
-    /// shutdown channel that signals the worker to exit on drop.
-    ///
-    /// `sender` is `None` when the cpal pipeline isn't running
-    /// (`AETHER_AUDIO_DISABLE=1`, no audio device, init failure). In
-    /// that mode `NoteOn` / `NoteOff` no-op and `SetMasterGain` replies
-    /// `Err`.
-    ///
-    /// Issue 629 / Phase B: `thread` and `shutdown` are
-    /// plain fields. Pre-Phase-A they sat behind a `Mutex<AudioTeardown>`
-    /// so `Drop::drop(&mut self)` could `.take()` them while handlers
-    /// ran with `&self` (Arc-shared). Post-Phase-A the dispatcher owns
-    /// the cap as `Box<A>` and `Drop` runs with exclusive `&mut self`,
-    /// so the wrapping mutex retires.
-    /// A `play_track` request parked while its `aether.fs.read` is in
-    /// flight (ADR-0103 §2). Keyed in [`AudioCapability::pending_tracks`]
-    /// by the echoed `(namespace, path)` the `ReadResult` carries; the
-    /// original requester's reply route + the synth-side track key live
-    /// here until the bytes land.
-    struct PendingTrack {
-        /// The original `play_track` requester — the `PlayTrackResult`
-        /// reply routes here across the fs round-trip + decode.
-        source: Source,
-        /// The synth-side track key's sender component, baked into the
-        /// `TrackStart` event so the lane keys by `(sender, lane,
-        /// namespace, path)` while the fs correlation keys by
-        /// `(namespace, path)`.
-        sender_mailbox: MailboxId,
-        /// The caller-supplied lane that disambiguates senders sharing a
-        /// source mailbox; part of the synth-side track key.
-        lane: Option<String>,
-        gain: f32,
-        looping: bool,
-    }
-
-    /// Completion context the `play_track` decode dispatch carries so the
-    /// `#[handler(task)]` arm can build the `TrackStart` event + the reply
-    /// without re-deriving anything (ADR-0093 §5). The worker produces the
-    /// decoded PCM; this carries the synth key + play parameters alongside.
-    struct TrackDecodeContext {
-        sender_mailbox: MailboxId,
-        lane: Option<String>,
-        namespace: String,
-        path: String,
-        gain: f32,
-        looping: bool,
-    }
-
-    /// Output of the decode dispatch worker — the resampled mono PCM, or
-    /// the decode failure to relay as `PlayTrackResult::Err`.
-    type DecodeOutput = Result<Vec<f32>, DecodeError>;
-
-    /// A `load_instrument` request parked while its `.sfz` `aether.fs.read`
-    /// is in flight (ADR-0103 §2/§5). Keyed in
-    /// [`AudioCapability::pending_instruments`] by the echoed
-    /// `(namespace, path)` of the `.sfz`. Only the original requester's
-    /// reply route lives here — the namespace / path come back on the
-    /// `ReadResult`, and the bank's name is derived from the `.sfz` path.
-    struct PendingInstrument {
-        source: Source,
-    }
-
-    /// One unique sample a bank assembly is fetching: the path as written
-    /// in the `.sfz` (resolved against `default_path`), the fs path it is
-    /// read from (joined with the `.sfz`'s own directory), and its bytes
-    /// once the `aether.fs.read` lands.
-    struct SampleSlot {
-        sample_rel: String,
-        fs_path: String,
-        bytes: Option<Vec<u8>>,
-    }
-
-    /// A bank load in progress: the `.sfz` parsed into regions, fanning out
-    /// one `aether.fs.read` per unique referenced sample, assembling when
-    /// the last reply lands (ADR-0103 §2). Keyed in
-    /// [`AudioCapability::assemblies`] by a minted id; the per-sample reads
-    /// correlate back to it through [`AudioCapability::pending_samples`].
-    struct BankAssembly {
-        /// The original `load_instrument` requester — the
-        /// `LoadInstrumentResult` reply routes here.
-        source: Source,
-        /// The fs namespace the `.sfz` and its samples live in (shared).
-        namespace: String,
-        /// The `.sfz` path — echoed on an `Err` reply for correlation.
-        sfz_path: String,
-        /// Bank name, derived from the `.sfz` filename stem.
-        name: String,
-        /// The parsed regions; each names a `sample_rel` resolved at
-        /// assembly time to its decoded PCM.
-        regions: Vec<SfzRegion>,
-        /// The unique samples, fetched in parallel.
-        samples: Vec<SampleSlot>,
-        /// How many samples are still missing their bytes; the bank
-        /// assembles when this reaches zero.
-        remaining: usize,
-    }
-
-    /// Completion context the bank-assembly dispatch carries so the
-    /// `#[handler(task)]` arm can build the `Err` reply (`Ok` carries the
-    /// assembled bank's own name / id / bytes). Mirrors
-    /// [`TrackDecodeContext`] for the load path.
-    struct BankAssemblyContext {
-        namespace: String,
-        path: String,
-    }
-
-    /// Output of the bank-assembly dispatch worker — the assembled,
-    /// device-rate bank behind an `Arc`, or a human-readable decode failure
-    /// to relay as `LoadInstrumentResult::Err`.
-    type BankAssemblyOutput = Result<Arc<SampleBank>, String>;
 
     pub struct AudioCapability {
         sender: Option<AudioEventSender>,
@@ -2491,144 +521,6 @@ mod native {
             }
             self.pending_samples.retain(|_, queue| !queue.is_empty());
         }
-    }
-
-    /// Decode every unique sample to device-rate mono PCM and assemble the
-    /// bank (ADR-0103 §6). Pure + `Send` so it runs on the blocking-dispatch
-    /// worker, off the realtime path. A failed decode aborts with a
-    /// human-readable reason the cap relays as `LoadInstrumentResult::Err`.
-    fn assemble_bank(
-        name: String,
-        regions: &[SfzRegion],
-        sample_bytes: &[(String, Vec<u8>)],
-        target_rate: u32,
-    ) -> BankAssemblyOutput {
-        // Decode each unique sample, carrying its source rate so loop frame
-        // offsets can be scaled by the same resample ratio applied to the
-        // PCM (ADR-0103 §6).
-        let mut decoded: Vec<(String, Arc<[f32]>, u32)> = Vec::with_capacity(sample_bytes.len());
-        let mut resident_bytes = 0usize;
-        for (rel, bytes) in sample_bytes {
-            let pcm =
-                decode_wav_to_mono(bytes, target_rate).map_err(|e| format!("sample {rel}: {e}"))?;
-            let source_rate = wav_source_rate(bytes).map_err(|e| format!("sample {rel}: {e}"))?;
-            resident_bytes += pcm.len() * size_of::<f32>();
-            decoded.push((rel.clone(), Arc::from(pcm.as_slice()), source_rate));
-        }
-
-        let mut bank_regions = Vec::with_capacity(regions.len());
-        for region in regions {
-            let (pcm, source_rate) = decoded
-                .iter()
-                .find(|(rel, _, _)| rel == &region.sample)
-                .map(|(_, pcm, source_rate)| (Arc::clone(pcm), *source_rate))
-                .ok_or_else(|| format!("region references unfetched sample {}", region.sample))?;
-            let loop_region = region
-                .loop_spec
-                .and_then(|lp| scale_loop(lp, source_rate, target_rate, pcm.len()));
-            bank_regions.push(SampleRegion {
-                lokey: region.lokey,
-                hikey: region.hikey,
-                lovel: region.lovel,
-                hivel: region.hivel,
-                pitch_keycenter: region.pitch_keycenter,
-                pcm,
-                loop_region,
-            });
-        }
-
-        Ok(Arc::new(SampleBank {
-            name,
-            regions: bank_regions,
-            resident_bytes,
-        }))
-    }
-
-    /// Read a WAV asset's source sample rate from its header (ADR-0103 §6).
-    /// Bank assembly needs it to scale a region's loop frame offsets by the
-    /// load-time resample ratio; `decode_wav_to_mono` consumes the same
-    /// header but only returns the resampled PCM. Parses the header chunk
-    /// only — the sample data is not read.
-    fn wav_source_rate(bytes: &[u8]) -> Result<u32, String> {
-        let reader = hound::WavReader::new(Cursor::new(bytes)).map_err(|e| e.to_string())?;
-        let rate = reader.spec().sample_rate;
-        if rate == 0 {
-            return Err("zero sample rate".to_owned());
-        }
-        Ok(rate)
-    }
-
-    /// Scale a region's source-frame loop bounds into device-rate fractional
-    /// positions (ADR-0103 §6): multiply by the resample ratio
-    /// `target_rate / source_rate` — the same ratio the PCM was resampled
-    /// by at load — and clamp `loop_end` to the resampled length. Returns
-    /// `None` when the resampled region is too short to loop or the bounds
-    /// collapse after clamping, degrading the region to unlooped.
-    #[allow(clippy::cast_precision_loss, clippy::cast_possible_truncation)]
-    fn scale_loop(
-        lp: SfzLoop,
-        source_rate: u32,
-        target_rate: u32,
-        resampled_len: usize,
-    ) -> Option<SampleLoop> {
-        if resampled_len < 2 || source_rate == 0 {
-            return None;
-        }
-        let ratio = f64::from(target_rate) / f64::from(source_rate);
-        let start = f64::from(lp.start) * ratio;
-        let end = (f64::from(lp.end) * ratio).min(resampled_len as f64);
-        if start + 1.0 >= end {
-            return None;
-        }
-        Some(SampleLoop {
-            start: start as f32,
-            end: end as f32,
-        })
-    }
-
-    /// The directory portion of an fs path (everything before the last
-    /// `/`), or `""` when the path has no directory. A bank's samples are
-    /// addressed relative to the `.sfz`'s own directory (ADR-0103 §5).
-    fn sfz_dir(path: &str) -> &str {
-        match path.rsplit_once('/') {
-            Some((dir, _)) => dir,
-            None => "",
-        }
-    }
-
-    /// Join a sample path onto the `.sfz`'s directory. An empty directory
-    /// leaves the sample as-is.
-    fn join_fs(dir: &str, rel: &str) -> String {
-        if dir.is_empty() {
-            rel.to_owned()
-        } else {
-            format!("{dir}/{rel}")
-        }
-    }
-
-    /// Derive a bank name from the `.sfz` filename stem (the last path
-    /// segment without its extension). Falls back to `"instrument"` for a
-    /// pathological empty stem.
-    fn bank_name_from_path(path: &str) -> String {
-        let file = path.rsplit('/').next().unwrap_or(path);
-        let stem = file.rsplit_once('.').map_or(file, |(stem, _)| stem);
-        if stem.is_empty() {
-            "instrument".to_owned()
-        } else {
-            stem.to_owned()
-        }
-    }
-
-    /// The first instrument id available to a loaded bank — one past the
-    /// last compiled-in built-in. The cap's `next_instrument_id` starts
-    /// here, the synth's bank table begins at the same offset.
-    fn builtin_id_ceiling() -> u8 {
-        // `BUILTINS` is a small fixed table (11 today); the length fits a
-        // `u8` with room to spare, and a load count that overflowed `u8`
-        // would be absurd.
-        #[allow(clippy::cast_possible_truncation)]
-        let n = BUILTINS.len() as u8;
-        n
     }
 
     impl Drop for AudioCapability {
